프로필 삭제

+++ GameSnap/frontend/src/axios/UserProfileAxios.js
@@ -64,7 +64,8 @@ export const profileAPI = {
 
   // 프로필 이미지 삭제
   deleteImage: async (userId) => {
-    return await apiClient.delete(`/profile/deleteImg/${userId}`);
+    console.log("memberId 삭제 이미지 보내기 전 :",userId);
+    return await apiClient.delete("/profile/deleteImg",{ params : {userId : userId}});
   },
 
   // 프로필 내용 업데이트
